springboot-crud-rest: exception handling I
in the description of the last commit, I mentioned that code was working fine for the happy scenarioes. but as soon as we were entering the wrong value, our system was facing  an exception and it was breaking the system. and showing us the 500 internal system error. And this is not a good way to handle the wrong input.

So to handle that, I added a condition here to make sure proper response must be displayed. And if that specific record does not exist in the system, then we show a custom json error. For that I created a class named "Student Error Response" and created a custom Exception named "StudentNotFoundException".  So now when we enter a record that does not exist in our system, then instead of showing a 500 error, we are showing a proper 404 error in json format.

things to learn:

1) ExceptionHandler annotation shows that this method is responsible for handling the exception. and the parameter of this function shows what kind of exceptions this method can handle. And in this case, it would be SutdentNotFoundException.

2) Eventhough we have handled the scenario in which record does not exist in the system, but there are some other cases, which our system is not properly handling right now. like if we enter a character, it won't handle it properly. I am handle these edge cases in next commit